refactor: tokenSafetyLookup
